+++ include/sound/fx.hpp
@@ -70,7 +70,7 @@ enum soundPriorityType {
 };
 
 struct smartSoundChannel {
-    long                whichSound;
+    int32_t             whichSound;
     int32_t             soundAge;
     short               soundVolume;
     soundPriorityType   soundPriority;
@@ -92,14 +92,14 @@ void ResetAllSounds();
 void PlayVolumeSound(
         short whichSoundID, uint8_t amplitude, short persistence, soundPriorityType priority);
 void PlayLocalizedSound(
-        unsigned long sx, unsigned long sy, unsigned long dx, unsigned long dy,
+        uint32_t sx, uint32_t sy, uint32_t dx, uint32_t dy,
         Fixed hvel, Fixed vvel, short whichSoundID, short amplitude,
         short persistence, soundPriorityType priority);
 void quiet_all();
 void SoundFXCleanup();
 
 void mPlayDistanceSound(
-        long mvolume, spaceObjectType* mobjectptr, long msoundid, long msoundpersistence,
+        int32_t mvolume, spaceObjectType* mobjectptr, int32_t msoundid, int32_t msoundpersistence,
         soundPriorityType msoundpriority);
 
 }  // namespace antares
